requirements.txt
chromadb>=0.5.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
rich>=13.0.0
python-frontmatter>=1.1.0
tiktoken>=0.7.0
